refactor(ngrx/store): move ngrx files to app/shared/stores
