+++ src/app/pages/notes/note-details/note-details.component.ts
@@ -3,8 +3,8 @@ import { Store, select } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
 import { INote } from '@app/shared/models/markdown-state.model';
-import * as fromRoot from '@app/pages/notes/shared/reducers';
-import { NotesActions } from '../shared/actions';
+import * as fromRoot from '@app/shared/store/reducers';
+import { NotesActions } from '../../../shared/store/actions';
 
 @Component({
   selector: 'app-note-details',
