+++ src/app/shared/components/card/card.component.ts
@@ -3,9 +3,9 @@ import { Store, select } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
 import { INote } from '../../models/markdown-state.model';
-import { NotesActions } from '@app/pages/notes/shared/actions';
-import { NavigationsActions } from '@app/pages/notes/shared/actions';
-import * as fromRoot from '@app/pages/notes/shared/reducers';
+import { NotesActions } from '@app/shared/store/actions';
+import { NavigationsActions } from '@app/shared/store/actions';
+import * as fromRoot from '@app/shared/store/reducers';
 
 @Component({
   selector: 'app-card',
