--- src/app/shared/enums/toolbars.enum.ts
@@ -1,3 +1,0 @@
-export enum Toolbar {
-  Preview,
-}
