+++ src/app/shared/store/effects/notes.effects.ts
@@ -5,11 +5,8 @@ import { of, EMPTY } from 'rxjs';
 import { Store, select } from '@ngrx/store';
 import { Update } from '@ngrx/entity';
 
-import { NotesService } from '../../notes.service';
-import {
-  NotesActions,
-  NavigationsActions,
-} from '@app/pages/notes/shared/actions';
+import { NotesService } from '../../../pages/notes/notes.service';
+import { NotesActions, NavigationsActions } from '@app/shared/store/actions';
 import * as fromRoot from '../reducers';
 import { MarkdownParser } from '@app/shared/services/markdown-parser/markdown-parser';
 import { INote } from '@app/shared/models/markdown-state.model';
