+++ src/app/shared/store/reducers/navigations.reducer.ts
