+++ src/app/shared/store/reducers/notes.reducer.ts
